requirements.txt
# Agent Framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0

# RAG and NLP
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# Web Interface
streamlit==1.28.0
streamlit-chat==0.1.1

# Data Processing
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0

# LLM APIs
openai>=1.0.0
anthropic>=0.7.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
requests>=2.25.0

# Development Tools
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.0